requirements.txt
# FloMind Dashboard Tresorerie -- dependances Python
# Genere le 27/04/2026 -- Python 3.12

streamlit==1.56.0
plotly==6.7.0
pandas==3.0.2
numpy==2.4.4
openpyxl==3.1.5
scikit-learn==1.8.0
bcrypt==4.2.1
fpdf2==2.7.9
pytest==9.0.3
pytest-cov==6.1.0

# Optionnel V1.5 -- connexion SQL
# pyodbc==5.1.0
# firebird-driver==1.10.6

# Optionnel V1.5 -- Google Sheets bridge
# gspread==6.1.3
# google-auth==2.38.0
